Tests/LibWeb/Text/expected/ShadowDOM/remove-style-element-inside-shadow-tree.txt

[object StyleSheetList]
Before remove, sheet count: 1
After remove, sheet count: 0
After setting innerText of removed sheet, we're still alive!
